chore: moved zoeken model classes a bit
