+++ src/main/java/net/atos/zac/app/zoeken/model/AbstractRESTZoekObject.java
@@ -5,7 +5,7 @@
 
 package net.atos.zac.app.zoeken.model;
 
-import net.atos.zac.zoeken.model.index.ZoekObjectType;
+import net.atos.zac.zoeken.model.zoekobject.ZoekObjectType;
 
 public abstract class AbstractRESTZoekObject {
