+++ src/main/java/net/atos/zac/app/zoeken/model/RESTZoekParameters.java
@@ -12,7 +12,7 @@ import net.atos.zac.zoeken.model.DatumVeld;
 import net.atos.zac.zoeken.model.FilterParameters;
 import net.atos.zac.zoeken.model.FilterVeld;
 import net.atos.zac.zoeken.model.SorteerVeld;
-import net.atos.zac.zoeken.model.index.ZoekObjectType;
+import net.atos.zac.zoeken.model.zoekobject.ZoekObjectType;
 
 public class RESTZoekParameters extends RestPageParameters {
