+++ src/main/java/net/atos/zac/solr/SolrSchemaUpdate.java
@@ -11,7 +11,7 @@ import java.util.Set;
 
 import org.apache.solr.client.solrj.request.schema.SchemaRequest;
 
-import net.atos.zac.zoeken.model.index.ZoekObjectType;
+import net.atos.zac.zoeken.model.zoekobject.ZoekObjectType;
 
 public interface SolrSchemaUpdate {
