+++ src/main/kotlin/net/atos/zac/signalering/ZaakTaskDueDateEmailNotificationService.kt
@@ -28,8 +28,8 @@ import net.atos.zac.zoeken.model.DatumVeld
 import net.atos.zac.zoeken.model.FilterVeld
 import net.atos.zac.zoeken.model.FilterWaarde
 import net.atos.zac.zoeken.model.ZoekParameters
-import net.atos.zac.zoeken.model.index.ZoekObjectType
 import net.atos.zac.zoeken.model.zoekobject.ZaakZoekObject
+import net.atos.zac.zoeken.model.zoekobject.ZoekObjectType
 import nl.lifely.zac.util.AllOpen
 import nl.lifely.zac.util.NoArgConstructor
 import org.flowable.task.api.Task
