+++ src/main/kotlin/net/atos/zac/zoeken/model/zoekobject/ZoekObject.kt
@@ -2,9 +2,7 @@
  * SPDX-FileCopyrightText: 2021 Atos, 2024 Lifely
  * SPDX-License-Identifier: EUPL-1.2+
  */
-package net.atos.zac.zoeken.model
-
-import net.atos.zac.zoeken.model.index.ZoekObjectType
+package net.atos.zac.zoeken.model.zoekobject
 
 interface ZoekObject {
     companion object {
